checking for candidate model existence

+++ code/05_candidate_models.R
@@ -8,7 +8,8 @@ rstan_options(auto_write=T)
 source("code/00_fn.R")
 
 # Full dataset
-ensFull_df <- read_csv("out/valid_df.csv")
+ensFull_df <- read_csv("out/valid_df.csv") |>
+  mutate(across(starts_with("sim_"), ~.x - mean(.x), .names="c_{.col}"))
 years <- unique(ensFull_df$year)
 
 sim_i <- read_csv("out/sim_2019-2023/sim_i.csv") |>
@@ -16,7 +17,6 @@ sim_i <- read_csv("out/sim_2019-2023/sim_i.csv") |>
          lab_short=if_else(fixDepth, "2D", "3D")) |>
   group_by(lab_short) |>
   mutate(lab=paste0(lab_short, ".", row_number())) 
-
 
 
 # individual simulation models --------------------------------------------
@@ -37,11 +37,15 @@ for(yr in seq_along(years)) {
     dat_rstan <- train_df |>
       select(-any_of(sim_i$sim[-i]), -any_of(paste0("c_", sim_i$sim[-i]))) |>
       make_data_rstan()
-    out_sim <- stan(file="code/stan/candidate_model.stan",
-                    model_name=glue("{sim}-{years[yr]}"), data=dat_rstan,
-                    chains=6, cores=6,iter=3000, warmup=2500,
-                    pars=c("b_b0", "b_IP", "Intercept_hu", "b_hu", "sigma"))
-    saveRDS(out_sim, glue("out/candidates/{sim}_CV-{years[yr]}_stanfit.rds"))
+    if(file.exists(glue("out/candidates/{sim}_CV-{years[yr]}_stanfit.rds"))) {
+      out_sim <- readRDS(glue("out/candidates/{sim}_CV-{years[yr]}_stanfit.rds"))
+    } else {
+      out_sim <- stan(file="code/stan/candidate_model.stan",
+                      model_name=glue("{sim}-{years[yr]}"), data=dat_rstan,
+                      chains=6, cores=6,iter=3000, warmup=2500,
+                      pars=c("b_b0", "b_IP", "Intercept_hu", "b_hu", "sigma"))
+      saveRDS(out_sim, glue("out/candidates/{sim}_CV-{years[yr]}_stanfit.rds"))
+    }
     CV_yr[[i]] <- make_predictions_candidate(out_sim, test_df, sim) |>
       colMeans() |>
       as_tibble() |>
@@ -75,11 +79,15 @@ for(yr in seq_along(years)) {
     dat_rstan <- train_df |>
       select(-any_of(sim_avgs[-i]), -any_of(paste0("c_", sim_avgs[-i]))) |>
       make_data_rstan()
+    if(file.exists(glue("out/ensembles/{sim}_CV-{years[yr]}_stanfit.rds"))) {
+      out_sim <- readRDS(glue("out/ensembles/{sim}_CV-{years[yr]}_stanfit.rds"))
+    } else {
     out_sim <- stan(file="code/stan/candidate_model.stan",
                     model_name=glue("{sim}-{years[yr]}"), data=dat_rstan,
                     chains=6, cores=6,iter=3000, warmup=2500,
                     pars=c("b_b0", "b_IP", "Intercept_hu", "b_hu", "sigma"))
     saveRDS(out_sim, glue("out/ensembles/{sim}_CV-{years[yr]}_stanfit.rds"))
+    }
     CV_yr[[i]] <- make_predictions_candidate(out_sim, test_df, sim) |>
       colMeans() |>
       as_tibble() |>
